labs/06-image-classification/python/image-clasification/requirements.txt
azure-core
azure-cognitiveservices-vision-customvision
python-dotenv
